# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service for find most freq words",
        "title": "Post checker",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/v1",
    "paths": {
        "/post/{post_id}/comments/statistics": {
            "get": {
                "description": "Get post statistic for most frequently posts.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Post"
                ],
                "summary": "Get post statistic",
                "operationId": "post_statistic",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id of specific post",
                        "name": "post_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.wordResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.Word": {
            "type": "object",
            "properties": {
                "post+id": {
                    "type": "string"
                },
                "word": {
                    "type": "string"
                }
            }
        },
        "v1.response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                }
            }
        },
        "v1.wordResponse": {
            "type": "object",
            "properties": {
                "word": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Word"
                    }
                }
            }
        }
    }
}